+++ servers/Azure.Mcp.Server/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 The Azure MCP Server updates automatically by default whenever a new release comes out 🚀. We ship updates twice a week on Tuesdays and Thursdays 😊
 
+## 0.5.11 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 0.5.10 (2025-08-28)
+
+### Bugs fixed
+
+- Fixed a bug with telemetry collection related to AppConfig tools. [[#44](https://github.com/microsoft/mcp/pull/44)]
+
 ## 0.5.9 (2025-08-26)
 
 ### Other Changes